README
Sitemaps, Userflow & Wireframes!

What is a sitemap?

A sitemap provides a diagram of the site’s hierarchy and shows where each page fits into this hierarchy.

What is a userflow?

User flow is the path taken by a prototypical user on a website or app to complete a task.

What is a wireframe?

A wireframe is a skeletal layout of a web page that demonstrates what interface elements will exist on key pages. 

Tools for user flows and wireframing:

- UxD
- Invision
- Figma
- Sketch

Homework:

Design & prototype 1. an interesting wireframe and userflow inspired by Net.Art. 2. the wireframe for a redesigned Patron portal